requirements.txt
django-leaflet>=0.22,<0.23
djangorestframework-gis>=0.12.0,<0.13.0
channels>=1.1.8,<1.2
openwisp-utils>=0.1.2,<0.2.0
